feat: add creation of Datasets via API
do not allow it though for everyone but only for token authenticated
requests
we need to create CategorySerializer, LicenseSerializer and
CitySerializer because we need to write and read from it.
`to_internal_value(self, data)` is needed when creating a model via
another model; maybe we could just use Charfields as we did with the
modeldataset and modelsubject, but this works for now.