enable per_page query param
to get all of the modeldatasets we want to change the per_page query param
we need to enable it first for it to work
and now we need to remove the page_size entry in the response because it would be not correct if the query param is set. no need to add it anyway.